+++ rover_drive_train/launch/drive_train.launch.py
@@ -1,0 +1,36 @@
+import os
+
+from ament_index_python.packages import get_package_share_directory
+
+from launch import LaunchDescription
+from launch.actions import DeclareLaunchArgument
+from launch.actions import IncludeLaunchDescription
+from launch.launch_description_sources import PythonLaunchDescriptionSource
+from launch.substitutions import LaunchConfiguration
+from launch.substitutions import TextSubstitution
+from launch_ros.actions import Node
+
+def generate_launch_description():
+        ld: LaunchDescription = LaunchDescription()
+
+        joy_launch_file = IncludeLaunchDescription(
+                PythonLaunchDescriptionSource(
+                        os.path.join(
+                                get_package_share_directory('rover_joy'),
+                                'launch/joy.launch.py'))
+        )
+
+        teleop_node= Node (package="rover_drive_train",
+                                    namespace="/main",
+                                    executable="teleop",
+                                    name="teleop",
+                                    parameters=[{"speed_factor_crawler": 0.01},
+                                                {"speed_factor_normal": 0.25},
+                                                {"speed_factor_turbo": 1.0},
+                                                {"smallest_radius": 0.30}]
+                                                )
+        
+        ld.add_action(joy_launch_file)
+        ld.add_action(teleop_node)
+
+        return ld
